Move ItemObject to new items_module.py
