+++ src/creatures/humanoids.py
@@ -5,6 +5,7 @@ import pygame as pg
 from pygame.math import Vector2
 from pygame.sprite import Group
 
+import items_module
 import model as mdl
 import mods
 
@@ -104,7 +105,7 @@ class Humanoid(mdl.DynamicObject):
 
         return called_ability
 
-    def attempt_pickup(self, item: mods.ItemObject) -> None:
+    def attempt_pickup(self, item: items_module.ItemObject) -> None:
         if item.mod.loc not in self.active_mods:
             self.equip(item.mod)
             item.kill()
