+++ src/items/bullet_weapons.py
@@ -3,7 +3,8 @@ from pygame.math import Vector2
 
 import images
 from data.mods_io import load_mod_data
-from mods import ItemObject, Mod
+from mods import Mod
+from items_module import ItemObject
 from projectiles import MuzzleFlash
