+++ src/items/rocks.py
@@ -6,7 +6,8 @@ from pygame.math import Vector2
 import images
 from data.mods_io import load_mod_data
 
-from mods import ItemObject, Mod
+from mods import Mod
+from items_module import ItemObject
 
 
 class RockObject(ItemObject):
