+++ src/items/utility_items.py
@@ -4,7 +4,8 @@ from pygame.math import Vector2
 import images
 
 from data.mods_io import load_mod_data
-from mods import ItemObject, Mod
+from mods import Mod
+from items_module import ItemObject
 
 
 class HealthPackObject(ItemObject):
